Functioning Environment
To use it all you have to do is pass the action through e.g. [1,0], [-1,0], [0,1], [0,-1] into the step function! 

This needs change where the action passed in will be an integer 0-3 that corresponds with a movement direction, 

Returns the Map, CurrentLocation, TargetLocation and whether or not the bot has arrived at the destination as observations.

Quite a bit of this still needs cleaned up (Not using built in observation/action spaces) but it at least works for basic search algorithms!

I have added an example file (CampusMapRun.py) which shows how to run the environment with an agent interacting with it. Again not perfect but an example that works!